+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this repository will be documented in this file.
 
+## Unreleased 
+
+### Fixed
+- Replaced ibopen calls with read_file 
+
 ## [1.0.1] - 2022-03-14
 
 ### Fixed